Change some internals = better performance
Use ScintillaWindow which is more efficient for making calls. Make "editor" just a reference. It is either "editor1" or "editor2". Previously it was its own Lua userdata. This also alleviates the need to request the current Scintilla handle with every call to "editor.XXX"

Quick testing shows this is almost 3 times faster in some cases